+++ Assets/PiRhoUtilities/Samples/PiRhoUtilitySamples/Scripts/ControlsSample.cs
@@ -25,6 +25,9 @@ namespace PiRhoSoft.Utilities
 		[ScenePicker]
 		public AssetReference Scene;
 
+		[AssetPicker("Sprites")]
+		public AssetReference Asset;
+
 		[Popup(new int[] { 0, 1, 2, 3, 4 }, new string[] { "Zero", "One", "Two", "Three", "Four" })]
 		public int IntPopup;